+++ modules/admin/views/articles/index.php
@@ -71,8 +71,11 @@ $this->title = 'Статьи';
         });
 
         $('.btn-delete').click(function () {
-            var $this = $(this);
-            deleteAjax($this, '<?php echo Url::toRoute('articles/delete'); ?>');
+            var agree = confirm('Вы действительно хотите удалить этот пункт?');
+            if (agree) {
+                var $this = $(this);
+                deleteAjax($this, '<?php echo Url::toRoute('articles/delete'); ?>');
+            }
         });
     });
 </script>
