+++ modules/admin/views/callback/index.php
@@ -81,8 +81,11 @@ $this->title = 'Обратный звонок';
         });
 
         $('.btn-delete').click(function () {
-            var $this = $(this);
-            deleteAjax($this, '<?php echo Url::toRoute('callback/delete'); ?>');
+            var agree = confirm('Вы действительно хотите удалить этот пункт?');
+            if (agree) {
+                var $this = $(this);
+                deleteAjax($this, '<?php echo Url::toRoute('callback/delete'); ?>');
+            }
         });
     });
 </script>
