+++ modules/admin/views/card/index.php
@@ -79,8 +79,11 @@ $this->title = 'Дисконтная карта';
         });
 
         $('.btn-delete').click(function () {
-            var $this = $(this);
-            deleteAjax($this, '<?php echo Url::toRoute('card/delete'); ?>');
+            var agree = confirm('Вы действительно хотите удалить этот пункт?');
+            if (agree) {
+                var $this = $(this);
+                deleteAjax($this, '<?php echo Url::toRoute('card/delete'); ?>');
+            }
         });
     });
 </script>
