+++ modules/admin/views/certificates/index.php
@@ -73,8 +73,11 @@ $this->title = 'Список';
         });
 
         $('.btn-delete').click(function () {
-            var $this = $(this);
-            deleteAjax($this, '<?php echo Url::toRoute('certificates/delete'); ?>');
+            var agree = confirm('Вы действительно хотите удалить этот пункт?');
+            if (agree) {
+                var $this = $(this);
+                deleteAjax($this, '<?php echo Url::toRoute('certificates/delete'); ?>');
+            }
         });
 
         $('.certificate').submit(function() {
