+++ modules/admin/views/opinions/index.php
@@ -63,8 +63,11 @@ $this->title = 'Список';
         });
 
         $('.btn-delete').click(function () {
-            var $this = $(this);
-            deleteAjax($this, '<?php echo Url::toRoute('opinions/delete'); ?>');
+            var agree = confirm('Вы действительно хотите удалить этот пункт?');
+            if (agree) {
+                var $this = $(this);
+                deleteAjax($this, '<?php echo Url::toRoute('opinions/delete'); ?>');
+            }
         });
     });
 </script>
