+++ modules/admin/views/prices/index.php
@@ -86,8 +86,11 @@ $this->title = 'Услуги';
         });
 
         $('.btn-delete').click(function () {
-            var $this = $(this);
-            deleteAjax($this, '<?php echo Url::toRoute('prices/delete'); ?>');
+            var agree = confirm('Вы действительно хотите удалить этот пункт?');
+            if (agree) {
+                var $this = $(this);
+                deleteAjax($this, '<?php echo Url::toRoute('prices/delete'); ?>');
+            }
         });
 
         $('#save-sort').click(function() {
